pyproject.toml
[project]
name = "uv-django-test"
version = "0.1.0"
description = "Django REST Framework project with PostgreSQL"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.0",
    "djangorestframework>=3.15.0",
    "django-rest-knox>=4.2.0",
    "django-cors-headers>=4.4.0",
    "python-decouple>=3.8",
    "whitenoise>=6.7.0",
    "dj-database-url>=2.2.0",
    "django-filter>=24.3.0",
    "pillow>=10.4.0",
    "cryptography>=45.0.4",
    "psycopg[binary,pool]>=3.2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-django>=4.9.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "pre-commit>=4.0.0",
]
prod = ["gunicorn>=23.0.0"]


[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "migrations",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.isort]
known-first-party = ["config", "apps"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = ["--strict-markers", "--strict-config", "--tb=short"]
